Add default suggestions to dictcc
It seems likely that the user wants to look up a word at point, or an
active region, with a higher frequency than usual.  Thus, add a default
suggestion that activates on an empty input, selecting for exactly these
circumstances.

The advantage to this solution over `dictcc-at-point` is that it is all
in one binding and it leaves the freedom for the user to change their
mind later in the process of looking up the word (i.e., only when
actually entering the query).