Kill selected region when inserting a candidate
When inserting a translation it often seems more intuitive to kill the
currently selected region, if any.

For example, if there is an active region around "force", we look up a
translation for it and then would like to insert "forcieren", replacing
the original word seems like the natural thing to do.

Related: https://github.com/martenlienen/dictcc.el/issues/14