Add completing-read backend
This is a new backend utilising the build-in completing-read interface.
So far, unlike the helm and ivy variants, one is only able to insert the
dictcc-source-lang version of the word.  This may seemingly be solved by
using a transient keymap, but that would force us to either

    a) have the user make that decision as the first keypress, or
    b) incorporate all of the keys invoking self-insert-command into the
       transient map,

neither of which seem like good options.